Make io testing channel usable across library boundaries
